Add SudoFile function and use it with root file system access is needed.
