+++ scripts/buildRpm.sh
@@ -38,7 +38,7 @@ cp packaging/$PACKAGE.spec ~/rpmbuild/SPECS
 rm -f $PACKAGE-$VERSION.tar.gz
 tar czf $PACKAGE-$VERSION.tar.gz ./*
 mkdir -p ~/rpmbuild/SOURCES
-cp $PACKAGE-$VERSION.tar.gz ~/rpmbuild/SOURCES
+mv $PACKAGE-$VERSION.tar.gz ~/rpmbuild/SOURCES
 cd ~/rpmbuild
